+++ testinfra/modules/package.py
@@ -31,11 +31,11 @@ class Package(Module):
 
         - apk (Alpine)
         - apt (Debian, Ubuntu, ...)
-        - pacman (Arch)
+        - pacman (Arch, Manjaro )
         - pkg (FreeBSD)
         - pkg_info (NetBSD)
         - pkg_info (OpenBSD)
-        - rpm (RHEL, Centos, Fedora, ...)
+        - rpm (RHEL, RockyLinux, Fedora, ...)
         """
         raise NotImplementedError
 
@@ -73,12 +73,15 @@ class Package(Module):
         if host.system_info.distribution and (
             host.system_info.distribution.lower()
             in (
+                "almalinux",
                 "centos",
+                "cloudlinux",
                 "fedora",
+                "rocky",
             )
         ):
             return RpmPackage
-        if host.system_info.distribution == "arch":
+        if host.system_info.distribution in ("arch", "manjarolinux"):
             return ArchPackage
         if host.exists("apk"):
             return AlpinePackage